Add files via upload
This Project helps in detecting whether a person has worn a face mask or not. The basic principle behind the project is to detect the Face and then  Nose of a Person, if we gather the coordinates of the Nose (which means the person is either wearing the mass improperly or else not wearing it) or if we don't get the coordinates of those it states that the person is wearing mask Properly.