Update README.md
The basic goal of this project is to detect whether a person is wearing a Face mask or not. The principle behind the python code is that if we detect the face of a person and the coordinates of its nose. If we are able to detect the coordinates of the Nose which means that either the person is not wearing a Face mask or the person is Face Mask improperly and if we are not able to detect the coordinates of a Nose which seems to be the Nose is covered hence the person has worn Face mask.